requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
Pillow>=8.0.0
Flask>=2.0.0
Flask-CORS>=3.0.0
numpy>=1.21.0
ultralytics>=8.0.0 